Move expressCare state to new reducer and action files
