+++ src/applications/vaos/actions/appointments.js
@@ -58,12 +58,6 @@ export const FETCH_REQUEST_MESSAGES_FAILED =
 export const FETCH_REQUEST_MESSAGES_SUCCEEDED =
   'vaos/FETCH_REQUEST_MESSAGES_SUCCEEDED';
 
-export const FETCH_EXPRESS_CARE_WINDOWS = 'vaos/FETCH_EXPRESS_CARE_WINDOWS';
-export const FETCH_EXPRESS_CARE_WINDOWS_FAILED =
-  'vaos/FETCH_EXPRESS_CARE_WINDOWS_FAILED';
-export const FETCH_EXPRESS_CARE_WINDOWS_SUCCEEDED =
-  'vaos/FETCH_EXPRESS_CARE_WINDOWS_SUCCEEDED';
-
 export const CANCEL_APPOINTMENT = 'vaos/CANCEL_APPOINTMENT';
 export const CANCEL_APPOINTMENT_CONFIRMED = 'vaos/CANCEL_APPOINTMENT_CONFIRMED';
 export const CANCEL_APPOINTMENT_CONFIRMED_SUCCEEDED =
@@ -233,66 +227,6 @@ export function fetchPastAppointments(startDate, endDate, selectedIndex) {
       dispatch({
         type: FETCH_PAST_APPOINTMENTS_FAILED,
         error,
-      });
-    }
-  };
-}
-
-export function fetchExpressCareWindows() {
-  return async (dispatch, getState) => {
-    dispatch({
-      type: FETCH_EXPRESS_CARE_WINDOWS,
-    });
-
-    const initialState = getState();
-    const appointments = initialState.appointments;
-    let parentFacilities = appointments.parentFacilities;
-    const userSiteIds = selectSystemIds(initialState);
-
-    try {
-      if (!parentFacilities) {
-        parentFacilities = await getOrganizations({
-          siteIds: userSiteIds,
-          useVSP: false,
-        });
-        if (parentFacilities.length) {
-          const ids = parentFacilities.map(parent => parent.id);
-          const facilityData = [];
-
-          if (ids.length < 20) {
-            const paramsArray = parentFacilities.map(parent => {
-              const rootOrg = getRootOrganization(parentFacilities, parent.id);
-              return {
-                siteId: getSiteIdFromOrganization(rootOrg || parent),
-                parentId: parent.id.replace('var', ''),
-                typeOfCareId: EXPRESS_CARE,
-              };
-            });
-
-            facilityData.push(
-              ...(await Promise.all(
-                paramsArray.map(p =>
-                  getFacilitiesBySystemAndTypeOfCare(
-                    p.siteId,
-                    p.parentId,
-                    p.typeOfCareId,
-                  ),
-                ),
-              )),
-            );
-          }
-
-          dispatch({
-            type: FETCH_EXPRESS_CARE_WINDOWS_SUCCEEDED,
-            facilityData,
-            nowUtc: moment.utc(),
-          });
-        }
-      }
-    } catch (error) {
-      captureError(error);
-      dispatch({
-        type: FETCH_EXPRESS_CARE_WINDOWS_FAILED,
       });
     }
   };
