+++ src/applications/vaos/reducers/index.js
@@ -1,5 +1,7 @@
 import appointments from './appointments';
+import expressCare from './expressCare';
 
 export default {
   appointments,
+  expressCare,
 };
